backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.7.0
python-dotenv==1.0.1
requests==2.32.3
